Fix post-merge issues: clippy warnings and doc tests
- Fixed DatabaseBuilder duplicate Default implementation
- Fixed PI constant approximation warnings
- Fixed 25+ clippy warnings in examples
- Fixed 17 failing doc tests in src/lib.rs
- All tests now pass, ready for CI

+++ examples/basic/simple_device.rs
@@ -91,7 +91,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     // Unconfirmed request (Who-Is)
     let whois_apdu = Apdu::UnconfirmedRequest {
-        service_choice: UnconfirmedServiceChoice::WhoIs,
+        service_choice: UnconfirmedServiceChoice::WhoIs as u8,
         service_data: whois_buffer,
     };
 
@@ -101,7 +101,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let decoded_apdu = Apdu::decode(&encoded_apdu)?;
     match decoded_apdu {
         Apdu::UnconfirmedRequest { service_choice, .. } => {
-            println!("Decoded APDU service choice: {}", service_choice as u8);
+            println!("Decoded APDU service choice: {}", service_choice);
         }
         _ => println!("Unexpected APDU type"),
     }
